# perf: Don't create new sets when checking compatibility